Move sorting out of utils directory
